requirements.txt
requests==2.22.0
python-telegram-bot==12.2.0
python-dotenv==0.10.3